# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WorldRAG API
  version: 1.0.0
  description: 端到端混合检索 RAG 知识库问答系统 API 契约。作者：晨星
servers:
  - url: http://localhost:8000

paths:
  /api/health:
    get:
      summary: 健康检查
      operationId: health
      responses:
        "200":
          description: 服务状态与活跃后端
          content:
            application/json:
              schema:
                type: object
                required: [status, chunks, backends]
                properties:
                  status: { type: string, enum: [ok] }
                  chunks: { type: integer, minimum: 0 }
                  backends:
                    type: object
                    additionalProperties: { type: string }

  /api/ingest:
    post:
      summary: 摄取文档（文本或文件路径）
      operationId: ingest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                texts:
                  type: array
                  items:
                    type: object
                    required: [text]
                    properties:
                      text: { type: string, minLength: 1 }
                      doc_id: { type: string }
                      metadata: { type: object }
                paths:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: 摄取结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested_chunks: { type: integer }
                  total_chunks: { type: integer }
        "400":
          description: 空请求或不支持的文件
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/query:
    post:
      summary: 检索问答
      operationId: query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string, minLength: 1 }
      responses:
        "200":
          description: 答案与引用
          content:
            application/json:
              schema:
                type: object
                required: [query, answer, citations]
                properties:
                  query: { type: string }
                  answer: { type: string }
                  citations:
                    type: array
                    items: { $ref: "#/components/schemas/Citation" }
                  backends:
                    type: object
                    additionalProperties: { type: string }
        "409":
          description: 知识库为空
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: 请求体校验失败
        "502":
          description: 生成失败
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /api/evaluate:
    post:
      summary: 在独立索引上运行离线评测
      operationId: evaluate
      responses:
        "200":
          description: 评测指标
          content:
            application/json:
              schema:
                type: object
                properties:
                  metrics:
                    type: object
                    properties:
                      recall_at_k: { type: number }
                      mrr: { type: number }
                      keyword_hit_rate: { type: number }
                      queries: { type: integer }
                      chunks: { type: integer }
                  backends:
                    type: object
                    additionalProperties: { type: string }
                  degradations:
                    type: object
                    additionalProperties: { type: string }
                  per_query:
                    type: array
                    items:
                      type: object
                      properties:
                        query: { type: string }
                        hit: { type: boolean }
                        rank: { type: [integer, "null"] }
                        keyword_hit: { type: boolean }
                        retrieved:
                          type: array
                          items: { type: string }

  /api/degradations:
    get:
      summary: 后端降级报告
      operationId: degradations
      responses:
        "200":
          description: 降级信息（正常时应为空对象）
          content:
            application/json:
              schema:
                type: object
                properties:
                  degradations:
                    type: object
                    additionalProperties: { type: string }

components:
  schemas:
    Citation:
      type: object
      required: [n, doc_id, chunk_id, score, source, snippet]
      properties:
        n: { type: integer }
        doc_id: { type: string }
        chunk_id: { type: string }
        score: { type: number }
        source: { type: string, enum: [vector, bm25, fusion, rerank] }
        snippet: { type: string }
    Error:
      type: object
      properties:
        detail: { type: string }